refactor: Auth, Member Service 추상화
